+++ vite-template/staticsite-template/pageData.js
@@ -1,0 +1,20 @@
+const baseData = {};
+
+const pageDate = {
+  "/index.html": {
+    meta: {
+      isHome: true,
+      title: "indexページだよ",
+      description: "indexページの説明文だよ",
+    },
+  },
+  "/about.html": {
+    meta: {
+      isHome: false,
+      title: "hogeページだよ",
+      description: "hogeページの説明文だよ",
+    },
+  },
+};
+
+export default pageDate;
