+++ vite-template/staticsite-template/src/about.html
@@ -3,9 +3,9 @@
   <head>
     {{> meta/head-meta }}
 
-    <link rel="stylesheet" href="./scss/style.scss" />
+    <link rel="stylesheet" href="./assets/scss/style.scss" />
 
-    <script src="./js/main.js" type="module"></script>
+    <script src="./assets/js/main.js" type="module"></script>
   </head>
   <body>
     <main class="l-main">
@@ -14,7 +14,7 @@
       <div class="p-test">projectテスト</div>
       <div class="c-test is-blue">componentテスト</div>
       <img
-        src="./img/100KB1.gif"
+        src="./aseets/img/100KB1.gif"
         alt=""
         width="640"
         height="640"
