+++ vite-template/staticsite-template/src/components/meta/head-meta.html
@@ -5,8 +5,8 @@
 <meta name="robots" content="noindex" />
 
 <!-- site setting -->
-<title>{{title}}</title>
-<meta name="description" content="{{description}}" />
+<title>{{meta.title}}</title>
+<meta name="description" content="{{meta.description}}" />
 <link rel="icon" href="/favicon.ico" />
 
 <!-- site ogp -->
