+++ vite-template/staticsite-template/src/index.html
@@ -10,25 +10,25 @@
   <body>
     <main class="l-main">
       {{> header }}
-      <div>INDEX</div>
-      <div class="p-test">projectテスト</div>
-      <div class="c-test">componentテスト</div>
+      <div style="font-size: 80px">Index Page</div>
+      <div class="p-test" style="margin: 20px 0">projectテスト</div>
+      <div class="c-test" style="margin: 20px 0">componentテスト</div>
       <img
-        src="./img/100KB1.gif"
+        src="./assets/img/100KB1.gif"
         alt=""
         width="640"
         height="640"
         style="width: 100px"
       />
       <img
-        src="./img/20602_color.png"
+        src="./assets/img/20602_color.png"
         alt=""
         width="670"
         height="642"
         style="width: 100px"
       />
       <img
-        src="./img/sunder-muthukumaran-fd6K_OFlnRA-unsplash.jpg"
+        src="./assets/img/sunder-muthukumaran-fd6K_OFlnRA-unsplash.jpg"
         alt=""
         width="640"
         height="640"
